requirements.txt
# Security Scanners
semgrep>=1.90.0

# AI/LLM Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# GitHub Integration
PyGithub>=2.4.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
tenacity>=9.0.0
pyyaml>=6.0.0

# Output Formatting
rich>=13.5.0,<14.0.0  # semgrep requires rich~=13.5.2
sarif-om>=1.0.4
jschema-to-python>=1.2.3

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
